requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
certifi==2022.5.18.1; python_version >= '3.6'
cffi==1.15.0
charset-normalizer==2.0.12; python_version >= '3.5'
cryptography==37.0.2
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-allauth==0.51.0
django-crispy-forms==1.14.0
django==3.2
idna==3.3; python_version >= '3.5'
oauthlib==3.2.0; python_version >= '3.6'
pillow==9.1.1
pycparser==2.21
pyjwt[crypto]==2.4.0; python_version >= '3.6'
python3-openid==3.2.0
pytube==12.1.0
pytz==2022.1
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.28.0; python_version >= '3.7' and python_version < '4'
sqlparse==0.4.2; python_version >= '3.5'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
